Gether ent logic in proto::single::ent
